@@ -8,35 +8,31 @@
     "frameworkHomeURL": "https://mobx.js.org/"
   },
   "scripts": {
-    "build-dev": "webpack --watch",
-    "build-prod": "webpack --mode production"
+    "build-dev": "rollup -c -w",
+    "build-prod": "rollup -c --environment production"
   },
-  "keywords": [
-    "react",
-    "webpack"
-  ],
-  "author": "Stefan Krause",
+  "author": "Maxim Molochkov",
   "license": "Apache-2.0",
   "homepage": "https://github.com/krausest/js-framework-benchmark",
   "repository": {
     "type": "git",
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
-  "devDependencies": {
-    "@babel/core": "7.21.0",
-    "@babel/plugin-proposal-class-properties": "7.16.7",
-    "@babel/plugin-proposal-decorators": "^7.21.0",
-    "@babel/preset-env": "7.20.2",
-    "@babel/preset-react": "7.18.6",
-    "babel-loader": "9.1.2",
-    "terser-webpack-plugin": "^5.3.7",
-    "webpack": "5.76.1",
-    "webpack-cli": "^5.0.1"
-  },
   "dependencies": {
-    "mobx": "5.15.4",
-    "mobx-react": "6.2.2",
-    "react": "17.0.1",
-    "react-dom": "17.0.1"
+    "mobx": "6.9.0",
+    "mobx-react-lite": "3.4.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "@babel/core": "7.21.8",
+    "@babel/preset-env": "7.21.5",
+    "@babel/preset-react": "7.18.6",
+    "@rollup/plugin-babel": "6.0.3",
+    "@rollup/plugin-commonjs": "25.0.0",
+    "@rollup/plugin-node-resolve": "15.0.2",
+    "@rollup/plugin-replace": "5.0.2",
+    "@rollup/plugin-terser": "0.4.3",
+    "rollup": "3.22.0"
   }
 }
